refactor: split app and ui god modules
- extract app responsibilities into focused modules
- split app tests by feature with shared helpers
- extract ui overlays and panels from ui mod orchestration
- keep ui mod as layout/render dispatcher
